Split hooks up into individual files
