+++ projects/packages/my-jetpack/_inc/components/my-jetpack-screen/index.jsx
@@ -31,7 +31,7 @@ import useProduct from '../../data/products/use-product';
 import useSimpleQuery from '../../data/use-simple-query';
 import { getMyJetpackWindowInitialState } from '../../data/utils/get-my-jetpack-window-state';
 import useAnalytics from '../../hooks/use-analytics';
-import useNotificationWatcher from '../../hooks/use-notification-watcher/use-notification-watcher';
+import useNotificationWatcher from '../../hooks/use-notification-watcher';
 import ConnectionsSection from '../connections-section';
 import IDCModal from '../idc-modal';
 import JetpackManageBanner from '../jetpack-manage-banner';
