+++ projects/packages/my-jetpack/_inc/hooks/use-notification-watcher/index.ts
@@ -1,0 +1,12 @@
+import { getMyJetpackWindowInitialState } from '../../data/utils/get-my-jetpack-window-state';
+import useBadInstallNotice from './use-bad-install-notice';
+import useSiteConnectionNotice from './use-site-connection-notice';
+
+const useNotificationWatcher = () => {
+	const { redBubbleAlerts } = getMyJetpackWindowInitialState();
+
+	useBadInstallNotice( redBubbleAlerts );
+	useSiteConnectionNotice( redBubbleAlerts );
+};
+
+export default useNotificationWatcher;
