+++ projects/packages/my-jetpack/_inc/hooks/use-notification-watcher/use-site-connection-notice.ts
@@ -1,24 +1,13 @@
 import { __, sprintf } from '@wordpress/i18n';
 import { useContext, useEffect } from 'react';
 import { MyJetpackRoutes } from '../../constants';
-import { NOTICE_PRIORITY_MEDIUM, NOTICE_PRIORITY_HIGH } from '../../context/constants';
+import { NOTICE_PRIORITY_HIGH } from '../../context/constants';
 import { NoticeContext } from '../../context/notices/noticeContext';
 import { useAllProducts } from '../../data/products/use-product';
-import { getMyJetpackWindowInitialState } from '../../data/utils/get-my-jetpack-window-state';
 import getProductSlugsThatRequireUserConnection from '../../data/utils/get-product-slugs-that-require-user-connection';
-import useAnalytics from '../use-analytics';
 import useMyJetpackNavigate from '../use-my-jetpack-navigate';
 
 type RedBubbleAlerts = Window[ 'myJetpackInitialState' ][ 'redBubbleAlerts' ];
-
-const useNotificationWatcher = () => {
-	const { redBubbleAlerts } = getMyJetpackWindowInitialState();
-
-	useBadInstallNotice( redBubbleAlerts );
-	useSiteConnectionNotice( redBubbleAlerts );
-};
-
-export default useNotificationWatcher;
 
 const useSiteConnectionNotice = ( redBubbleAlerts: RedBubbleAlerts ) => {
 	const { setNotice } = useContext( NoticeContext );
@@ -69,55 +58,4 @@ const useSiteConnectionNotice = ( redBubbleAlerts: RedBubbleAlerts ) => {
 	}, [ navToConnection, products, redBubbleAlerts, setNotice ] );
 };
 
-const useBadInstallNotice = ( redBubbleAlerts: RedBubbleAlerts ) => {
-	const { setNotice } = useContext( NoticeContext );
-	const { recordEvent } = useAnalytics();
-
-	useEffect( () => {
-		const badInstallAlerts = Object.keys( redBubbleAlerts ).filter( key =>
-			key.endsWith( '-bad-installation' )
-		) as Array< `${ string }-bad-installation` >;
-
-		if ( badInstallAlerts.length === 0 ) {
-			return;
-		}
-
-		const alert = redBubbleAlerts[ badInstallAlerts[ 0 ] ];
-		const { plugin } = alert.data;
-		const devEnvUrl =
-			'https://github.com/Automattic/jetpack/blob/trunk/docs/development-environment.md';
-
-		const errorMessage = sprintf(
-			// translators: %s is the name of the plugin that has a bad installation.
-			__(
-				'Your installation of %1$s is incomplete. If you installed %1$s from GitHub, please refer to the developer documentation to set up your development environment.',
-				'jetpack-my-jetpack'
-			),
-			plugin
-		);
-
-		const onCtaClick = () => {
-			window.open( devEnvUrl );
-			recordEvent( 'jetpack_my_jetpack_bad_installation_notice_cta_click', {
-				plugin,
-			} );
-		};
-
-		const noticeOptions = {
-			status: 'error',
-			actions: [
-				{
-					label: __( 'See documentation', 'jetpack-my-jetpack' ),
-					onClick: onCtaClick,
-					noDefaultClasses: true,
-				},
-			],
-		};
-
-		setNotice( {
-			message: errorMessage,
-			options: noticeOptions,
-			priority: NOTICE_PRIORITY_MEDIUM,
-		} );
-	}, [ redBubbleAlerts, setNotice, recordEvent ] );
-};
+export default useSiteConnectionNotice;
